Scenario_2_Vendor_Contract/agent_build/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "vendor-contract-review-agent"
version = "0.1.0"
description = "AI-assisted first-pass review of vendor contracts with legal playbook enforcement"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.0.0",
    "pydantic-core>=2.13.0",
    "python-multipart>=0.0.6",
    "PyPDF2>=4.0.0",
    "python-docx>=0.8.11",
    "sqlalchemy>=2.0.0",
    "anthropic>=0.20.0",
    "pyyaml>=6.0.0",
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "httpx>=0.25.0",
]

[project.optional-dependencies]
dev = [
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "mypy>=1.7.0",
]
